feat: add logging and annotating to controller
This super large commit handles the logging and annotating of new pods. Also the
helm charts have been stripped down to be just what they need as well as adding
rbac so the controller can do the things it needs. This adds the scripts to
create a local kind cluster (with registry) as well as tearing all resources
down.